Split off async class
